+++ lib/ltpa.mli
@@ -26,26 +26,6 @@ module Rsa : sig
     }
     [@@deriving eq,ord,show]
 
-    val bin_t : t Bin_prot.Type_class.t0
-    [@@ocaml.deprecated "Bin_prot serializers will be removed in key-parsers 1.0.0"]
-    val bin_size_t : t -> int
-    [@@ocaml.deprecated "Bin_prot serializers will be removed in key-parsers 1.0.0"]
-    val bin_read_t : Bin_prot.Common.buf -> pos_ref:Bin_prot.Common.pos_ref -> t
-    [@@ocaml.deprecated "Bin_prot serializers will be removed in key-parsers 1.0.0"]
-    val bin_write_t : Bin_prot.Common.buf -> pos:int -> t -> int
-    [@@ocaml.deprecated "Bin_prot serializers will be removed in key-parsers 1.0.0"]
-    val bin_shape_t : Bin_prot.Shape.t
-    [@@ocaml.deprecated "Bin_prot serializers will be removed in key-parsers 1.0.0"]
-    val bin_reader_t : t Bin_prot.Type_class.reader0
-    [@@ocaml.deprecated "Bin_prot serializers will be removed in key-parsers 1.0.0"]
-    val bin_writer_t : t Bin_prot.Type_class.writer0
-    [@@ocaml.deprecated "Bin_prot serializers will be removed in key-parsers 1.0.0"]
-
-    val to_yojson : t -> Yojson.Safe.json
-    [@@ocaml.deprecated "Yojson serializers will be removed in key-parsers 1.0.0"]
-    val of_yojson : Yojson.Safe.json -> (t, string) result
-    [@@ocaml.deprecated "Yojson serializers will be removed in key-parsers 1.0.0"]
-
     val decode : Cstruct.t -> (t, string) Result.result
   end
 
@@ -63,26 +43,6 @@ module Rsa : sig
       n: Z.t;
     }
     [@@deriving eq,ord,show]
-
-    val bin_t : t Bin_prot.Type_class.t0
-    [@@ocaml.deprecated "Bin_prot serializers will be removed in key-parsers 1.0.0"]
-    val bin_size_t : t -> int
-    [@@ocaml.deprecated "Bin_prot serializers will be removed in key-parsers 1.0.0"]
-    val bin_read_t : Bin_prot.Common.buf -> pos_ref:Bin_prot.Common.pos_ref -> t
-    [@@ocaml.deprecated "Bin_prot serializers will be removed in key-parsers 1.0.0"]
-    val bin_write_t : Bin_prot.Common.buf -> pos:int -> t -> int
-    [@@ocaml.deprecated "Bin_prot serializers will be removed in key-parsers 1.0.0"]
-    val bin_shape_t : Bin_prot.Shape.t
-    [@@ocaml.deprecated "Bin_prot serializers will be removed in key-parsers 1.0.0"]
-    val bin_reader_t : t Bin_prot.Type_class.reader0
-    [@@ocaml.deprecated "Bin_prot serializers will be removed in key-parsers 1.0.0"]
-    val bin_writer_t : t Bin_prot.Type_class.writer0
-    [@@ocaml.deprecated "Bin_prot serializers will be removed in key-parsers 1.0.0"]
-
-    val to_yojson : t -> Yojson.Safe.json
-    [@@ocaml.deprecated "Yojson serializers will be removed in key-parsers 1.0.0"]
-    val of_yojson : Yojson.Safe.json -> (t, string) result
-    [@@ocaml.deprecated "Yojson serializers will be removed in key-parsers 1.0.0"]
 
     val decode : Cstruct.t -> (t, string) Result.result
   end