+++ tests/test_all.ml
@@ -5,7 +5,6 @@ let suite =
   [ "LTPA" >::: Test_ltpa.suite
   ; "ASN1" >::: Test_asn1.suite
   ; "CVC" >::: Test_cvc.suite
-  ; "Derivable" >::: Test_derivable.suite
   ]
 
 let _ = run_test_tt_main suite